requirements.txt
aniso8601==9.0.1
asgiref==3.6.0
attrs==22.2.0
Django==4.1.5
exceptiongroup==1.1.0
graphene==3.2.1
graphene-django==3.0.0
graphql-core==3.2.3
graphql-relay==3.2.0
iniconfig==2.0.0
packaging==23.0
pluggy==1.0.0
promise==2.3
pytest==7.2.1
pytest-django==4.5.2
six==1.16.0
sqlparse==0.4.3
text-unidecode==1.3
tomli==2.0.1
